MenuActivity
Added an abstract class MenuActivity which holds the menu navigation interface and disables the back button.
